==> Preprocessing/4.4_Tissue_correction_filtered.Rmd <==
---
title: "Batch correction across tissues"
author: "Karol Nowicki-Osuch, based on Nils Eling"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float:
      toc_collapsed: true
    toc_depth: 3
    number_sections: false
    theme: paper
    code_folding: hide
knit: (function(inputFile, encoding) { rmarkdown::render(inputFile, 
        encoding = encoding, output_file = '~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Reports/4.4_Tissue_correction_filtered.html') })
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
# export RSTUDIO_PANDOC=/usr/lib/rstudio-server/bin/pandoc
# Rscript -e "library(rmarkdown); render('4.4_Tissue_correction_filtered.Rmd')"
knitr::opts_chunk$set(echo = TRUE)
```

# Read in the data

```{r libraries, message=FALSE}
library(scran)
library(scater)
library(ggplot2)
library(DropletUtils)
library(openxlsx)
library(Rtsne)
library(stats)
library(knitr)
source("../Analysis/Functions/auxiliary.R")
```

# Batch correction across tissues

I use data without patient 12 that might have gastric metaplasia samples.

```{r batch-correction, results='asis', }
sce.list <- readRDS("~/Dropbox/Postdoc/2019-12-29_BE2020/All_sce_filtered_high_quality.rds")
tissue.list <- list()

for(i in c("NE", "NG", "BE", "NSCJ", "BSCJ", "ND", "SMG")){
  
  cur_sces <- sce.list[which(grepl(i, names(sce.list)))]
  
  cur_sces <- lapply(cur_sces, function(n){
    rowData(n) <- rowData(n)[,1:2]
    n
  })
  
  #  cur_sces <-lapply(cur_sces, function (x) {
  #   if(colnames(rowData(x))[3] == "Type") {
  #     colnames(rowData(x))[3]<-"is_feature_control"
  #   } else {
  #     colnames(rowData(x))[3]<-colnames(rowData(x))[3]
  #   }
  #   x
  # })
  
  # Order sce object by size
  cur_sces <- cur_sces[order(unlist(lapply(cur_sces, ncol)), decreasing = TRUE)]
  
  
  
  # Combine sce objects
  tissue_sce <- do.call("cbind", cur_sces)
  qclusters <- quickCluster(tissue_sce)
  tissue_sce <- computeSumFactors(tissue_sce, clusters=qclusters)
  tissue_sce <- logNormCounts(tissue_sce)
  
  # Batch correction
  corrected <- batch.correction(cur_sces)
  
  # Save batch corrected counts in metdata
  metadata(tissue_sce)$corrected <- corrected
  
  # Compute tsne on corrected counts
  set.seed(1234)
  tsne <- Rtsne(t(corrected), pca = FALSE)
  
  # Store tsne in slot
  reducedDims(tissue_sce)$TSNE <- tsne$Y[,1:2]
  
  # Clustering on corrected data
  #dist.all <- as.dist(sqrt((1 - cor(as.matrix(corrected), 
  #                                    method = "spearman"))/2))
  #dendro <- hclust(dist.all, method = "ward.D2")
  #cluster <- as.character(cutreeDynamic(dendro = dendro, distM = as.matrix(dist.all), 
  #                      minClusterSize = 10, deepSplit = 0))
  if(i == "SMG") {
    g <- buildSNNGraph(corrected, k = 4, d = NA) # Use of k = 4 allows for capture of ductal cells as a seperate cluster in SMG samples. 
    cluster <- igraph::cluster_louvain(g)$membership
  } else {
    g <- buildSNNGraph(corrected, k = 5, d = NA) # A good balance needed to separate the samples.
    cluster <- igraph::cluster_louvain(g)$membership
  }

  # Save clustering in new slot
  colData(tissue_sce)$Tissue_cluster <- cluster
  
  # Visualize clustering
  cur_p <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                             tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                             clusters = as.factor(colData(tissue_sce)$Tissue_cluster))) +
    geom_point(aes(tSNE1, tSNE2, colour = clusters))+ggtitle(i)
  print(cur_p)
  
  
  cur_p <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                             tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                             clusters = as.factor(colData(tissue_sce)$Patient))) +
    geom_point(aes(tSNE1, tSNE2, colour = clusters))+ggtitle(i)
  print(cur_p)
  
  # Visualize whether populations are patient specific 
  print(kable(table(cluster, colData(tissue_sce)$Patient),row.names = TRUE))
  cat("\n")
  
  
  # Perform differential expression
  markers <- findMarkers(tissue_sce, groups = cluster, block = colData(tissue_sce)$Patient)
  
  
  markers.spec <- lapply(markers, function(n){
    if(!is.na(n$Top[1]) & !is.nan(sum(as.matrix(n[1,3:ncol(n)])))){
      test_n <- !is.na(n[1,3:ncol(n)])[1,]
      cur_n <- n[n$FDR < 0.1 & apply(n[,3:ncol(n)], 1, function(x){sum(x[test_n] > 0)}) == sum(test_n),]
      if(nrow(cur_n) > 0){
        cur_n$GeneName <- rowData(tissue_sce)$Symbol[match(rownames(cur_n), rowData(tissue_sce)$ID)]
      }
    }
    else{
      cur_n <- NULL
    }
    cur_n
  })
  
  if(!dir.exists(paste("~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Marker_genes/Tissues_filtered/", 
                       i, sep = ""))){
    dir.create(paste("~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Marker_genes/Tissues_filtered/", 
                     i, sep = ""))
  }
  
  write.xlsx(markers.spec, paste("~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Marker_genes/Tissues_filtered/", i, "/Marker_genes.xlsx", sep = ""))
  
  tissue.list[[i]] <- tissue_sce
}




saveRDS(tissue.list, "~/Dropbox/Postdoc/2019-12-29_BE2020/Tissue_sce_filtered.rds")
```

# End Matter

To finish get session info:

```{r Endnote, include=TRUE, echo=TRUE, warning=FALSE, eval=TRUE, message=FALSE, tidy=TRUE}
sessionInfo()
```
