==> Preprocessing/5.4_Reclustering.Rmd <==
---
title: "Batch correction across tissues"
author: "Karol Nowicki-Osuch, based on Nils Eling"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float:
      toc_collapsed: true
    toc_depth: 3
    number_sections: false
    theme: paper
    code_folding: hide
knit: (function(inputFile, encoding) { rmarkdown::render(inputFile, 
        encoding = encoding, output_file = '~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Reports/5.4_Reclustering.html') })
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
# export RSTUDIO_PANDOC=/usr/lib/rstudio-server/bin/pandoc
# Rscript -e "library(rmarkdown); render('4.4_Tissue_correction_filtered.Rmd')"
knitr::opts_chunk$set(echo = TRUE)
```

# Read in the data

```{r libraries, message=FALSE}
library(scran)
library(scater)
library(ggplot2)
library(DropletUtils)
library(openxlsx)
library(Rtsne)
library(stats)
library(knitr)
library(viridis)
source("../Analysis/Functions/auxiliary.R")
```

# NSCJ 

The NSCJ samples contains a small population of cells that expresses MUC5B and some of the cells have features of transitional epithelium (defined as KRT7+/TP63+/KRT5+). I would like to see if these cells can be reclustered into additional populations.

```{r, results='asis', }
sce.list <- readRDS("~/Dropbox/Postdoc/2019-12-29_BE2020/Tissue_sce_filtered.rds")
tissue_sce <- sce.list$NSCJ


set.seed(1234)

# Visualize clustering
cur_p <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                           tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                           clusters = as.factor(colData(tissue_sce)$Tissue_cluster))) +
  geom_point(aes(tSNE1, tSNE2, colour = clusters))
print(cur_p)


# Visualize expression
for (gene in c("MUC5B", "KRT7", "KRT5", "TP63", "KRT23")) {
  cur_exp <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                               tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                               clusters = as.factor(colData(tissue_sce)$Tissue_cluster),
                               expression = logcounts(tissue_sce)[rowData(tissue_sce)$Symbol == gene,])) +
    geom_point(aes(tSNE1, tSNE2, colour = expression)) +
    scale_colour_viridis(name = "log2(Expr)") + ggtitle(gene)
  
  print(cur_exp)
  
  cur_exp <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                               tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                               clusters = as.factor(colData(tissue_sce)$Tissue_cluster),
                               expression = logcounts(tissue_sce)[rowData(tissue_sce)$Symbol == gene,])) +
    geom_boxplot(aes(clusters, expression))  + ggtitle(gene)
  
  print(cur_exp)
}



# Visualize whether populations are patient specific 
print(kable(table(colData(tissue_sce)$Tissue_cluster, colData(tissue_sce)$Patient),row.names = TRUE))
cat("\n")
```

## Cluster 12 seem to be containing these cells

```{r}
cur_sce <- tissue_sce[,colData(tissue_sce)$Tissue_cluster == 12]

## 212 cells

# Plot the patient data
cur_p <- ggplot(data.frame(tSNE1 = reducedDims(cur_sce)$TSNE[,1],
                           tSNE2 = reducedDims(cur_sce)$TSNE[,2],
                           clusters = as.factor(colData(cur_sce)$Patient))) +
  geom_point(aes(tSNE1, tSNE2, colour = clusters))
print(cur_p)

# Extract corrected dimensions for the selected cells and perform clustering
corrected <- metadata(tissue_sce)$corrected[,colData(tissue_sce)$Tissue_cluster == 12]
g <- buildSNNGraph(corrected, k = 10, d = NA) # A good balance needed to separate the samples.
cluster <- igraph::cluster_louvain(g)$membership

# Visualise the clustering results
cur_p <- ggplot(data.frame(tSNE1 = reducedDims(cur_sce)$TSNE[,1],
                           tSNE2 = reducedDims(cur_sce)$TSNE[,2],
                           clusters = as.factor(cluster))) +
  geom_point(aes(tSNE1, tSNE2, colour = clusters))
print(cur_p)

# Visualize expression
for (gene in c("MUC5B", "KRT7", "KRT5", "TP63", "KRT23")) {
  cur_exp <- ggplot(data.frame(tSNE1 = reducedDims(cur_sce)$TSNE[,1],
                               tSNE2 = reducedDims(cur_sce)$TSNE[,2],
                               clusters = as.factor(cluster),
                               expression = logcounts(cur_sce)[rowData(cur_sce)$Symbol == gene,])) +
    geom_point(aes(tSNE1, tSNE2, colour = expression)) +
    scale_colour_viridis(name = "log2(Expr)") + ggtitle(gene)
  
  print(cur_exp)
  
  cur_exp <- ggplot(data.frame(tSNE1 = reducedDims(cur_sce)$TSNE[,1],
                               tSNE2 = reducedDims(cur_sce)$TSNE[,2],
                               clusters = as.factor(cluster),
                               expression = logcounts(cur_sce)[rowData(cur_sce)$Symbol == gene,])) +
    geom_boxplot(aes(clusters, expression))  + ggtitle(gene)
  
  print(cur_exp)
}

# Visualize whether populations are patient specific 
print(kable(table(cluster, colData(cur_sce)$Patient),row.names = TRUE))
cat("\n")

cluster[cluster == 1] <- 12
cluster[cluster == 2] <- 27
cluster[cluster == 3] <- 28
cluster[cluster == 4] <- 29

#rename the clusters in the original object
colData(tissue_sce)$Tissue_cluster[colData(tissue_sce)$Tissue_cluster == 12] <- cluster

# Visualize whether populations are patient specific 
print(kable(table(colData(tissue_sce)$Tissue_cluster, colData(tissue_sce)$Patient),row.names = TRUE))
cat("\n")

# Visualize clustering
cur_p <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                           tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                           clusters = as.factor(colData(tissue_sce)$Tissue_cluster))) +
  geom_point(aes(tSNE1, tSNE2, colour = clusters))
print(cur_p)

# Perform differential expression
markers <- findMarkers(tissue_sce, groups = colData(tissue_sce)$Tissue_cluster, block = colData(tissue_sce)$Patient)


markers.spec <- lapply(markers, function(n){
  if(!is.na(n$Top[1]) & !is.nan(sum(as.matrix(n[1,3:ncol(n)])))){
    test_n <- !is.na(n[1,3:ncol(n)])[1,]
    cur_n <- n[n$FDR < 0.1 & apply(n[,3:ncol(n)], 1, function(x){sum(x[test_n] > 0)}) == sum(test_n),]
    if(nrow(cur_n) > 0){
      cur_n$GeneName <- rowData(tissue_sce)$Symbol[match(rownames(cur_n), rowData(tissue_sce)$ID)]
    }
  }
  else{
    cur_n <- NULL
  }
  cur_n
})

if(!dir.exists(paste("~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Marker_genes/Tissues_filtered/", 
                     "NSCJ", sep = ""))){
  dir.create(paste("~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Marker_genes/Tissues_filtered/", 
                   "NSCJ", sep = ""))
}

write.xlsx(markers.spec, paste("~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Marker_genes/Tissues_filtered/", "NSCJ", "/Marker_genes_reclustered.xlsx", sep = ""))


sce.list$NSCJ <- tissue_sce

```

# NSCJ - Endocrine
The NSCJ samples contains a small population of cells that expresses CHGA and GHRL and they are not split into two cluster. I will fix it here.

```{r, results='asis', }
tissue_sce <- sce.list$NSCJ


set.seed(1234)

# Visualize clustering
cur_p <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                           tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                           clusters = as.factor(colData(tissue_sce)$Tissue_cluster))) +
  geom_point(aes(tSNE1, tSNE2, colour = clusters))
print(cur_p)


# Visualize expression
for (gene in c("GHRL", "CHGA")) {
  cur_exp <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                               tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                               clusters = as.factor(colData(tissue_sce)$Tissue_cluster),
                               expression = logcounts(tissue_sce)[rowData(tissue_sce)$Symbol == gene,])) +
    geom_point(aes(tSNE1, tSNE2, colour = expression)) +
    scale_colour_viridis(name = "log2(Expr)") + ggtitle(gene)
  
  print(cur_exp)
  
  cur_exp <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                               tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                               clusters = as.factor(colData(tissue_sce)$Tissue_cluster),
                               expression = logcounts(tissue_sce)[rowData(tissue_sce)$Symbol == gene,])) +
    geom_boxplot(aes(clusters, expression))  + ggtitle(gene)
  
  print(cur_exp)
}



# Visualize whether populations are patient specific 
print(kable(table(colData(tissue_sce)$Tissue_cluster, colData(tissue_sce)$Patient),row.names = TRUE))
cat("\n")
```

## Cluster 26 seem to be containing endocrine cells

```{r}
cur_sce <- tissue_sce[,colData(tissue_sce)$Tissue_cluster == 26]

## 147 cells

# Plot the patient data
cur_p <- ggplot(data.frame(tSNE1 = reducedDims(cur_sce)$TSNE[,1],
                           tSNE2 = reducedDims(cur_sce)$TSNE[,2],
                           clusters = as.factor(colData(cur_sce)$Patient))) +
  geom_point(aes(tSNE1, tSNE2, colour = clusters))
print(cur_p)

# Extract corrected dimensions for the selected cells and perform clustering
corrected <- metadata(tissue_sce)$corrected[,colData(tissue_sce)$Tissue_cluster == 26]
g <- buildSNNGraph(corrected, k = 50, d = NA) # A good balance needed to separate the samples.
cluster <- igraph::cluster_louvain(g)$membership

# Visualise the clustering results
cur_p <- ggplot(data.frame(tSNE1 = reducedDims(cur_sce)$TSNE[,1],
                           tSNE2 = reducedDims(cur_sce)$TSNE[,2],
                           clusters = as.factor(cluster))) +
  geom_point(aes(tSNE1, tSNE2, colour = clusters))
print(cur_p)

# Visualize expression
for (gene in c("GHRL", "CHGA")) {
  cur_exp <- ggplot(data.frame(tSNE1 = reducedDims(cur_sce)$TSNE[,1],
                               tSNE2 = reducedDims(cur_sce)$TSNE[,2],
                               clusters = as.factor(cluster),
                               expression = logcounts(cur_sce)[rowData(cur_sce)$Symbol == gene,])) +
    geom_point(aes(tSNE1, tSNE2, colour = expression)) +
    scale_colour_viridis(name = "log2(Expr)") + ggtitle(gene)
  
  print(cur_exp)
  
  cur_exp <- ggplot(data.frame(tSNE1 = reducedDims(cur_sce)$TSNE[,1],
                               tSNE2 = reducedDims(cur_sce)$TSNE[,2],
                               clusters = as.factor(cluster),
                               expression = logcounts(cur_sce)[rowData(cur_sce)$Symbol == gene,])) +
    geom_boxplot(aes(clusters, expression))  + ggtitle(gene)
  
  print(cur_exp)
}

# Visualize whether populations are patient specific 
print(kable(table(cluster, colData(cur_sce)$Patient),row.names = TRUE))
cat("\n")

cluster[cluster == 1] <- 26
cluster[cluster == 2] <- 30

#rename the clusters in the original object
colData(tissue_sce)$Tissue_cluster[colData(tissue_sce)$Tissue_cluster == 26] <- cluster

# Visualize whether populations are patient specific 
print(kable(table(colData(tissue_sce)$Tissue_cluster, colData(tissue_sce)$Patient),row.names = TRUE))
cat("\n")

# Visualize clustering
cur_p <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                           tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                           clusters = as.factor(colData(tissue_sce)$Tissue_cluster))) +
  geom_point(aes(tSNE1, tSNE2, colour = clusters))
print(cur_p)

# Perform differential expression
markers <- findMarkers(tissue_sce, groups = colData(tissue_sce)$Tissue_cluster, block = colData(tissue_sce)$Patient)


markers.spec <- lapply(markers, function(n){
  if(!is.na(n$Top[1]) & !is.nan(sum(as.matrix(n[1,3:ncol(n)])))){
    test_n <- !is.na(n[1,3:ncol(n)])[1,]
    cur_n <- n[n$FDR < 0.1 & apply(n[,3:ncol(n)], 1, function(x){sum(x[test_n] > 0)}) == sum(test_n),]
    if(nrow(cur_n) > 0){
      cur_n$GeneName <- rowData(tissue_sce)$Symbol[match(rownames(cur_n), rowData(tissue_sce)$ID)]
    }
  }
  else{
    cur_n <- NULL
  }
  cur_n
})

if(!dir.exists(paste("~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Marker_genes/Tissues_filtered/", 
                     "NSCJ", sep = ""))){
  dir.create(paste("~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Marker_genes/Tissues_filtered/", 
                   "NSCJ", sep = ""))
}

write.xlsx(markers.spec, paste("~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Marker_genes/Tissues_filtered/", "NSCJ", "/Marker_genes_reclustered.xlsx", sep = ""))


sce.list$NSCJ <- tissue_sce

```


# BSCJ 

The BSCJ samples contains a small population of cells that expresses MUC2 and CHGA. These are markers of goblet cells and endocrine cells and they are expressed in different cell populations. I would like to separate them into two clusters.

```{r, results='asis', }
tissue_sce <- sce.list$BSCJ



# Visualize clustering
cur_p <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                           tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                           clusters = as.factor(colData(tissue_sce)$Tissue_cluster))) +
  geom_point(aes(tSNE1, tSNE2, colour = clusters))
print(cur_p)

# Visualize expression
for (gene in c("MUC2", "SPINK4", "CHGA")) {
  cur_exp <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                               tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                               clusters = as.factor(colData(tissue_sce)$Tissue_cluster),
                               expression = logcounts(tissue_sce)[rowData(tissue_sce)$Symbol == gene,])) +
    geom_point(aes(tSNE1, tSNE2, colour = expression)) +
    scale_colour_viridis(name = "log2(Expr)") + ggtitle(gene)
  
  print(cur_exp)
  
  cur_exp <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                               tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                               clusters = as.factor(colData(tissue_sce)$Tissue_cluster),
                               expression = logcounts(tissue_sce)[rowData(tissue_sce)$Symbol == gene,])) +
    geom_boxplot(aes(clusters, expression))  + ggtitle(gene)
  
  print(cur_exp)
}
# cluster 16



# Visualize whether populations are patient specific 
print(kable(table(colData(tissue_sce)$Tissue_cluster, colData(tissue_sce)$Patient),row.names = TRUE))
cat("\n")
```

## Cluster 16 seem to be containing these cells

```{r}
cur_sce <- tissue_sce[,colData(tissue_sce)$Tissue_cluster == 16]

## 70 cells

# Plot the patient data
cur_p <- ggplot(data.frame(tSNE1 = reducedDims(cur_sce)$TSNE[,1],
                           tSNE2 = reducedDims(cur_sce)$TSNE[,2],
                           clusters = as.factor(colData(cur_sce)$Patient))) +
  geom_point(aes(tSNE1, tSNE2, colour = clusters))
print(cur_p)

# Extract corrected dimensions for the selected cells and perform clustering
corrected <- metadata(tissue_sce)$corrected[,colData(tissue_sce)$Tissue_cluster == 16]
g <- buildSNNGraph(corrected, k = 10, d = NA) # A good balance needed to separate the samples.
cluster <- igraph::cluster_louvain(g)$membership

# Visualise the clustering results
cur_p <- ggplot(data.frame(tSNE1 = reducedDims(cur_sce)$TSNE[,1],
                           tSNE2 = reducedDims(cur_sce)$TSNE[,2],
                           clusters = as.factor(cluster))) +
  geom_point(aes(tSNE1, tSNE2, colour = clusters))
print(cur_p)

# Visualize expression
for (gene in c("MUC2", "SPINK4", "CHGA")) {
  cur_exp <- ggplot(data.frame(tSNE1 = reducedDims(cur_sce)$TSNE[,1],
                               tSNE2 = reducedDims(cur_sce)$TSNE[,2],
                               clusters = as.factor(cluster),
                               expression = logcounts(cur_sce)[rowData(cur_sce)$Symbol == gene,])) +
    geom_point(aes(tSNE1, tSNE2, colour = expression)) +
    scale_colour_viridis(name = "log2(Expr)") + ggtitle(gene)
  
  print(cur_exp)
  
  cur_exp <- ggplot(data.frame(tSNE1 = reducedDims(cur_sce)$TSNE[,1],
                               tSNE2 = reducedDims(cur_sce)$TSNE[,2],
                               clusters = as.factor(cluster),
                               expression = logcounts(cur_sce)[rowData(cur_sce)$Symbol == gene,])) +
    geom_boxplot(aes(clusters, expression))  + ggtitle(gene)
  
  print(cur_exp)
}


# Visualize whether populations are patient specific 
print(kable(table(cluster, colData(cur_sce)$Patient),row.names = TRUE))
cat("\n")

cluster[cluster == 1] <- 16
cluster[cluster == 2] <- (max(colData(tissue_sce)$Tissue_cluster) + 1)

#rename the clusters in the original object
colData(tissue_sce)$Tissue_cluster[colData(tissue_sce)$Tissue_cluster == 16] <- cluster

# Visualize whether populations are patient specific 
print(kable(table(colData(tissue_sce)$Tissue_cluster, colData(tissue_sce)$Patient),row.names = TRUE))
cat("\n")

# Visualize clustering
cur_p <- ggplot(data.frame(tSNE1 = reducedDims(tissue_sce)$TSNE[,1],
                           tSNE2 = reducedDims(tissue_sce)$TSNE[,2],
                           clusters = as.factor(colData(tissue_sce)$Tissue_cluster))) +
  geom_point(aes(tSNE1, tSNE2, colour = clusters))
print(cur_p)

# Perform differential expression
markers <- findMarkers(tissue_sce, groups = colData(tissue_sce)$Tissue_cluster, block = colData(tissue_sce)$Patient)


markers.spec <- lapply(markers, function(n){
  if(!is.na(n$Top[1]) & !is.nan(sum(as.matrix(n[1,3:ncol(n)])))){
    test_n <- !is.na(n[1,3:ncol(n)])[1,]
    cur_n <- n[n$FDR < 0.1 & apply(n[,3:ncol(n)], 1, function(x){sum(x[test_n] > 0)}) == sum(test_n),]
    if(nrow(cur_n) > 0){
      cur_n$GeneName <- rowData(tissue_sce)$Symbol[match(rownames(cur_n), rowData(tissue_sce)$ID)]
    }
  }
  else{
    cur_n <- NULL
  }
  cur_n
})

if(!dir.exists(paste("~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Marker_genes/Tissues_filtered/", 
                     "BSCJ", sep = ""))){
  dir.create(paste("~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Marker_genes/Tissues_filtered/", 
                   "BSCJ", sep = ""))
}

write.xlsx(markers.spec, paste("~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Marker_genes/Tissues_filtered/", "BSCJ", "/Marker_genes_reclustered.xlsx", sep = ""))


sce.list$BSCJ <- tissue_sce

saveRDS(sce.list, "~/Dropbox/Postdoc/2019-12-29_BE2020/Tissue_sce_filtered.rds")

```



# End Matter

To finish get session info:

```{r Endnote, include=TRUE, echo=TRUE, warning=FALSE, eval=TRUE, message=FALSE, tidy=TRUE}
sessionInfo()
```
