==> Preprocessing/6.4_Full_data_batch_correction_filtered.Rmd <==
---
title: "Batch correction across tissues and patients"
author: "Karol Nowicki-Osuch, based on Nils Eling"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float:
      toc_collapsed: true
    toc_depth: 3
    number_sections: false
    theme: paper
    code_folding: hide
knit: (function(inputFile, encoding) { rmarkdown::render(inputFile, 
        encoding = encoding, output_file = '~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Reports/6.4_Full_data_batch_correction_filtered.html') })
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
# Rscript -e "library(rmarkdown); render('6.4_Full_data_batch_correction_filtered.Rmd')"
knitr::opts_chunk$set(echo = TRUE)
```

# Read in the data

```{r libraries, message=FALSE}
library(scran)
library(scater)
library(DropletUtils)
library(openxlsx)
library(Rtsne)
library(umap)
library(RColorBrewer)
source("../Analysis/Functions/auxiliary.R")
```

# Batch correction across tissues

```{r batch-correction}
sce.list <- readRDS("~/Dropbox/Postdoc/2019-12-29_BE2020/All_sce_filtered_high_quality.rds")

sce.list <- lapply(sce.list, function(n){
  rowData(n) <- rowData(n)[,1:2]
  n
})


# Order sce objects for batch correction
sce.list <- sce.list[order(unlist(lapply(sce.list, ncol)), decreasing = TRUE)]

n <- names(sce.list)
sce.list <- sce.list[c(which(grepl("NSCJ", n)), which(grepl("BSCJ", n)), 
                       which(grepl("NE", n)), which(grepl("NG", n)),
                       which(grepl("BE", n)), which(grepl("ND", n)),
                       which(grepl("SMG", n)))]


# Combine sce objects
sce <- do.call("cbind", sce.list)

# Batch correction
corrected <- batch.correction(sce.list)

# Save batch corrected counts in metdata
metadata(sce)$corrected <- corrected

# Compute tsne on corrected counts
set.seed(1234)
tsne <- Rtsne(t(corrected), pca = FALSE)
umap <- umap(t(corrected))

# Store tsne in slot
reducedDims(sce)$TSNE <- tsne$Y[,1:2]
reducedDims(sce)$umap <- umap$layout[,1:2]

# Clustering on corrected data
g <- buildSNNGraph(corrected, k = 10)
clusters <- igraph::cluster_louvain(g)$membership

# Save clustering in new slot
colData(sce)$Global_cluster <- clusters

# Visualize clustering
ggplot(data.frame(tSNE1 = tsne$Y[,1],
                  tSNE2 = tsne$Y[,2],
                  clusters = as.factor(clusters))) + 
  geom_point(aes(tSNE1, tSNE2, colour = clusters), size = 0.5)



# Tissue
ggplot(data.frame(tSNE1 = tsne$Y[,1],
                  tSNE2 = tsne$Y[,2],
                  tissue = as.factor(colData(sce)$Tissue))) + 
  geom_point(aes(tSNE1, tSNE2, colour = tissue), size = 0.5) + 
  scale_color_brewer(palette = "Set1")

ggplot(data.frame(UMAP1 = umap$layout[,1],
                  UMAP2 = umap$layout[,2],
                  tissue = as.factor(colData(sce)$Tissue))) + 
  geom_point(aes(UMAP1, UMAP2, colour = tissue), size = 0.5) + 
  scale_color_brewer(palette = "Set1")

# Patient
ggplot(data.frame(tSNE1 = tsne$Y[,1],
                  tSNE2 = tsne$Y[,2],
                  patient = as.factor(colData(sce)$Patient))) + 
  geom_point(aes(tSNE1, tSNE2, colour = patient), size = 0.5) + 
  scale_color_manual(values = c(brewer.pal(8, "Set1"), brewer.pal(8, "Set3")))

ggplot(data.frame(UMAP1 = umap$layout[,1],
                  UMAP2 = umap$layout[,2],
                  patient = as.factor(colData(sce)$Patient))) + 
  geom_point(aes(UMAP1, UMAP2, colour = patient), size = 0.5) + 
  scale_color_manual(values = c(brewer.pal(8, "Set1"), brewer.pal(8, "Set3")))

# Patient and tissue
qual_col_pals = brewer.pal.info[brewer.pal.info$category == 'qual',]
col_vector = unlist(mapply(brewer.pal, qual_col_pals$maxcolors, rownames(qual_col_pals)))

ggplot(data.frame(tSNE1 = tsne$Y[,1],
                  tSNE2 = tsne$Y[,2],
                  all = as.factor(paste(colData(sce)$Tissue, colData(sce)$Patient)))) + 
  geom_point(aes(tSNE1, tSNE2, colour = all), size = 0.5) + 
  scale_color_manual(values = col_vector)

ggplot(data.frame(UMAP1 = umap$layout[,1],
                  UMAP2 = umap$layout[,2],
                  all = as.factor(paste(colData(sce)$Tissue, colData(sce)$Patient)))) + 
  geom_point(aes(UMAP1, UMAP2, colour = all), size = 0.5) + 
  scale_color_manual(values = col_vector)


# Plot all patients individually 
# Generate patient colour vector
patient_vector <- c(brewer.pal(8, "Set1"), brewer.pal(8, "Set3"))[1:length(unique(sce$Patient))]
names(patient_vector) <- unique(sce$Patient)

for(i in c("NE", "NSCJ", "BSCJ", "BE", "NG", "ND", "SMG")){
  cur_tsne <- tsne$Y[sce$Tissue == i,]
  cur_patient <- sce$Patient[sce$Tissue == i]
  
  print(ggplot(data.frame(tSNE1 = cur_tsne[,1],
                          tSNE2 = cur_tsne[,2],
                          patient = cur_patient)) + 
          geom_point(aes(tSNE1, tSNE2, colour = patient), size = 0.5) + 
          scale_color_manual(values = patient_vector) + ggtitle(i))
}

table(colData(sce)$Patient, colData(sce)$Tissue)

colSums(table(colData(sce)$Patient, colData(sce)$Tissue))


# Perform differential expression
markers <- findMarkers(sce, groups = colData(sce)$Global_cluster, 
                       block = paste(colData(sce)$Patient, colData(sce)$Tissue, sep = "_"))

markers.spec <- lapply(markers, function(n){
  if(!is.na(n$Top[1]) & !is.nan(sum(as.matrix(n[1,4:ncol(n)])))){
    test_n <- !is.na(n[1,4:ncol(n)])[1,]
    cur_n <- n[n$FDR < 0.1 & apply(n[,4:ncol(n)], 1, function(x){sum(x[test_n] > 0)}) == sum(test_n),]
    if(nrow(cur_n) > 0){
      cur_n$GeneName <- rowData(sce)$Symbol[match(rownames(cur_n), rowData(sce)$ID)]
    }
  }
  else{
    cur_n <- NULL
  }
  cur_n
})

write.xlsx(markers.spec, paste("~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Marker_genes/Global_filtered/Marker_genes.xlsx", sep = ""))

saveRDS(sce, "~/Dropbox/Postdoc/2019-12-29_BE2020/All_corrected_sce_filtered.rds")
```

# Add tissue clutsering to batch corrected data

```{r}
all_sce <- readRDS("~/Dropbox/Postdoc/2019-12-29_BE2020/All_corrected_sce_filtered.rds")
tissue_sce <- readRDS("~/Dropbox/Postdoc/2019-12-29_BE2020/Tissue_sce_filtered.rds")

tissue.clusters <- vector(length = ncol(all_sce))
names(tissue.clusters) <- paste(colData(all_sce)$Barcode, 
                                colData(all_sce)$Patient,
                                colData(all_sce)$Tissue, sep = "_")

for(i in 1:length(tissue_sce)){
  tissue.clusters[paste(colData(tissue_sce[[i]])$Barcode, 
                        colData(tissue_sce[[i]])$Patient,
                        colData(tissue_sce[[i]])$Tissue, sep = "_")] <- colData(tissue_sce[[i]])$Tissue_cluster
}

colData(all_sce)$Tissue_cluster <- tissue.clusters

saveRDS(all_sce, "~/Dropbox/Postdoc/2019-12-29_BE2020/All_corrected_sce_filtered.rds")
```

# End Matter

To finish get session info:

```{r Endnote, include=TRUE, echo=TRUE, warning=FALSE, eval=TRUE, message=FALSE, tidy=TRUE}
sessionInfo()
```
