==> Preprocessing/7.4_Cluster_annotation.Rmd <==
---
title: "Cluster annotation and colouring"
author: "Karol Nowicki-Osuch, based on Nils Eling code"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_depth: 2
knit: (function(inputFile, encoding) { rmarkdown::render(inputFile, 
        encoding = encoding, output_file = '~/Dropbox/Postdoc/2019-12-29_BE2020/Results/Reports/7.4_Cluster_annotation.html') })
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
# export RSTUDIO_PANDOC=/usr/lib/rstudio-server/bin/pandoc
# Rscript -e "library(rmarkdown); render('7.4_Cluster_annotation.Rmd')"
knitr::opts_chunk$set(echo = TRUE)
```

This script connects the cluster labels received from unsupervised, tissue-level clustering and links them to known cell types.

# Read in the data

```{r libraries, message=FALSE}
library(scran)
library(scater)
library(openxlsx)
library(Rtsne)
source("../Analysis/Functions/auxiliary.R")
sce <- readRDS("~/Dropbox/Postdoc/2019-12-29_BE2020/All_corrected_sce_filtered.rds")
```

# Annotation

Cell-type annotation needs to be done by hand.
We therefore create a .xlsx file that stores the cluster IDs and annotation.

```{r}
annot <- unique(paste(sce$Tissue, sce$Tissue_cluster, sep = "_"))
df.annot <- data.frame(tissue = sub("_[0-9]*$", "", annot),
                       cluster = sub("^[A-Z2]*_", "", annot), stringsAsFactors = FALSE)
df.annot <- df.annot[order(df.annot[,1], as.integer(df.annot[,2])),]
df.annot$cell_type <- vector(length = nrow(df.annot))
df.annot$include <- vector(length = nrow(df.annot))

write.xlsx(df.annot, "~/Dropbox/Postdoc/2019-12-29_BE2020/Additional_files/Cell_type_annotation.xlsx")
```

After cluster annotation, we will save this in the colData slot of the SCE object.

```{r}
annot <- read.xlsx("~/Dropbox/Postdoc/2019-12-29_BE2020/Additional_files/Cell_type_annotation_v10.xlsx")

# Link the tissue/cluster annotation of the single cell data to the annotation dataframe
m <- match(paste(sce$Tissue, sce$Tissue_cluster, sep = "_"), 
           paste(annot$tissue, annot$cluster, sep = "_"))
sce$cell_type <- annot$cell_type[m]
sce$cell_type_secondary <- annot$cell_type_secondary[m]
sce$tissue_type <- annot$Tissue_type[m]
sce$include <- annot$include[m]
sce$confidence <- annot$Confidence[m]
```

# Create colour vector for tissue samples

To allow consistent colouring of the tissues, we will store a colour vector in the metadata file.

```{r}
colour_vector <- vector(length = length(unique(sce$Tissue)))
names(colour_vector) <- unique(sce$Tissue)
colour_vector["NG"] <- "#4DBBD5FF"
colour_vector["BSCJ"] <- "#DC0000FF"
colour_vector["NE"] <- "darkred"
colour_vector["NSCJ"] <- "#F39B7FFF"
colour_vector["BE"] <- "#00A087FF"
colour_vector["ND"] <- "#3C5488FF"
colour_vector["SMG"] <- "#B09C85FF"

# Recompute tsne for visualization
set.seed(1234)
# tsne <- Rtsne(t(metadata(sce)$corrected), pca = FALSE, perplexity = 100)
  
# Store tsne in slot
# reducedDims(sce)$TSNE <- tsne$Y[,1:2]

ggplot(data.frame(tsne1 = reducedDims(sce)$TSNE[,1],
                  tsne2 = reducedDims(sce)$TSNE[,2],
                  tissue = sce$Tissue)) + 
  geom_point(aes(tsne1, tsne2, colour = tissue)) + 
  scale_colour_manual(values = colour_vector)

metadata(sce)$colour_vector <- colour_vector

saveRDS(sce, "~/Dropbox/Postdoc/2019-12-29_BE2020/All_corrected_sce_filtered.rds")
```

# End Matter

To finish get session info:

```{r Endnote, include=TRUE, echo=TRUE, warning=FALSE, eval=TRUE, message=FALSE, tidy=TRUE}
sessionInfo()
```
